Application.properties in gitignore, invalid token reduced

+++ src/main/java/com/example/DevOpsProj/controller/GitRepositoryController.java
@@ -1,17 +1,13 @@
 package com.example.DevOpsProj.controller;
 import com.example.DevOpsProj.commons.enumerations.EnumRole;
 import com.example.DevOpsProj.dto.responseDto.GitRepositoryDTO;
-import com.example.DevOpsProj.exceptions.NotFoundException;
 import com.example.DevOpsProj.model.GitRepository;
 import com.example.DevOpsProj.service.GitRepositoryService;
 import com.example.DevOpsProj.service.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -24,6 +20,8 @@ public class GitRepositoryController {
     @Autowired
     private JwtService jwtService;
 
+    private static final String INVALID_TOKEN = "Invalid Token";
+
     @PostMapping("/add")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<Object> createRepository(@RequestBody GitRepository gitRepository,
@@ -32,7 +30,7 @@ public class GitRepositoryController {
         if (isTokenValid) {
             return ResponseEntity.ok(gitRepositoryService.createRepository(gitRepository));
         } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid Token");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(INVALID_TOKEN);
         }
     }
 
@@ -42,7 +40,7 @@ public class GitRepositoryController {
         if (isTokenValid) {
             return ResponseEntity.ok(gitRepositoryService.getAllRepositories());
         } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid Token");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(INVALID_TOKEN);
         }
     }
 
@@ -55,7 +53,7 @@ public class GitRepositoryController {
             List<GitRepositoryDTO> gitRepositoryDTOS = gitRepositoryService.getAllRepositoriesByProject(id);
             return new ResponseEntity<>(gitRepositoryDTOS, HttpStatus.OK);
         } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid Token");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(INVALID_TOKEN);
         }
     }
 
@@ -68,7 +66,7 @@ public class GitRepositoryController {
             EnumRole enumRole = EnumRole.valueOf(role.toUpperCase());
             return ResponseEntity.ok(gitRepositoryService.getAllReposByRole(enumRole));
         } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid Token");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(INVALID_TOKEN);
         }
     }
 
@@ -86,27 +84,26 @@ public class GitRepositoryController {
                 return ResponseEntity.notFound().build();
             }
         } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid Token");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(INVALID_TOKEN);
         }
     }
-
 
     @DeleteMapping("/delete/{repoId}")
     public ResponseEntity<Object> deleteRepository(
             @PathVariable Long repoId,
             @RequestHeader(value = "AccessToken", required = false) String accessToken
     ) {
-        boolean isTokenValid = false;
-        if (accessToken != null && !accessToken.isEmpty()) {
-            isTokenValid = jwtService.isTokenTrue(accessToken);
-        }
+        boolean isTokenValid = jwtService.isTokenTrue(accessToken);
+        if(isTokenValid) {
+            try {
+                gitRepositoryService.deleteRepository(repoId); // Delete the repository
+                return ResponseEntity.ok("Deleted successfully");
 
-        try {
-            gitRepositoryService.deleteRepository(repoId); // Delete the repository
-            return ResponseEntity.ok("Deleted successfully");
-
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error");
+            } catch (Exception e) {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(INVALID_TOKEN);
         }
     }
 }
